+++ src/Model/Figure/figure.h
@@ -10,25 +10,25 @@ class Figure {
   void Clear();
 
   std::vector<double> getVertexes() const;
-  std::vector<size_t> getFacets() const;
-  size_t getCountVertexes() const;
-  size_t getCountEdges() const;
-  size_t getCountPolygons() const;
+  std::vector<int> getFacets() const;
+  int getCountVertexes() const;
+  int getCountEdges() const;
+  int getCountPolygons() const;
   double getMaxCoordinate() const;
 
   void setVertexes(std::vector<double>& vertexes);
-  void setFacets(std::vector<size_t>& facets);
-  void setCountVertexes(size_t countVertexes);
-  void setCountEdges(size_t countEdges);
-  void setCountPolygons(size_t countPolygons);
+  void setFacets(std::vector<int>& facets);
+  void setCountVertexes(int countVertexes);
+  void setCountEdges(int countEdges);
+  void setCountPolygons(int countPolygons);
   void setMaxCoordinate(double maxCoordinate);
 
  private:
   std::vector<double> vertexes_;
-  std::vector<size_t> facets_;
-  size_t countVertexes_;
-  size_t countEdges_;
-  size_t countPolygons_;
+  std::vector<int> facets_;
+  int countVertexes_;
+  int countEdges_;
+  int countPolygons_;
   double maxCoordinate_;
 };
 }  // namespace s21
